Add support for redux
While trying to do the integration with react-dnd on several lines, I realized that I had to pass the state through several components.
I did not find any example of how to handle react-dnd with that much layers without using redux and I suspect that it is because the application becomes too complex if we do.

I inspired myself alot from the redux documentation and example. Lucky enough, they use create-react-app for their example, so I simply used their template.

This commit breaks the support for Trello and the DnD, so it needs to be reimplemented in the following commits.

Among the features added:
- Create a card
- Delete a card
- A card can be active or completed independently from the list it belongs to (I don’t know if it’s a good idea though, we should revisit)
- A card can be edited
- A footer has been added and shows how many cards are active