Generate the lists and the cards from Trello board
This aims to reimplement the trello integration with the new Redux logic.
There is now two reducers, one for the cards and another one for the lists.

We now have a KanbanBoard component that takes an id. It will take care of the downloading of the cards and lists from Trello.
Eventually, this should become a container, but we will see if there is a need later.

The cards live independently from the the list. That means that if you add a new card, it won’t be displayed anywhere until you add it to a list.