Fix the create card component
Changing the create card behavior to have a default list instead of no list at all, this will be simplier for the development.
We default to the hardcoded To Do list for now, but it will certainly change.

Also, this is fixing the order of addition of the cards and lists, we now add a new one at the end of the array.