Implement the DnD logic for a single list
This uses an action called moveCard implemented in the cards reducer.
The tricky part is that we want to play with the index of the cards. As we have a reducer for the cards, the index will not be the index in the list, so we have to implement the logic somewhere else.
I tried to implement a single reducer for the lists and handle the logic like this, but the solution was less elegant, it would remove some flexibility in later stages.

I also remove the completed behavior for a card, we will implement it if needed later.