Adds synchronization with Trello
Before that, the trello integration was only unidirectionnal.
This is not the case anymore, we can add, edit and delete cards from Krdz.

The calls for the API are async, as there is no feature for synchronizing the functions (will come in ES7), we are having two actions: one that make the call to the API, and another one that will be called by the callback of the first one. This is inspired by redux async documentation, it seems to be a good pattern for now.