Adds a node backend that will service the static files.
The goal is to begin the work on having a more flexible application. Now that the Kaban view is advanced enough, it would be good to have some more pages, routing, authentification, all kind of stuff.
We use node as the backend, it was the easiest one to setup and the one recommended by the create-react-app documentation.
I followed this tutorial: https://www.fullstackreact.com/articles/using-create-react-app-with-a-server/ and used the files from the associated github account.
I also inspired myself from: https://medium.com/@patriciolpezjuri/using-create-react-app-with-react-router-express-js-8fa658bf892d#.g7i822ytn