Adds `build` directory to the source control
This is the way heroku will be able to serve the files.
This does not change anything for the development workflow, however, we will need to recompile the source code an push it to the source control every time we want to deploy.
This seems weird… Need to investigate it, there would be a way to automate it with Docker: https://blog.heroku.com/introducing_heroku_docker_release_build_deploy_heroku_apps_with_docker