Implements the routing system with two pages
We have a new Root component that will take care of the routing.
The App components does not display anything but takes care of a few configuration such as the error messages.
The old App component is now called Kanban (probably to be merged w/ the KanbanBoard component).

Also changing the types import module to be named ActionTypes in actions/index.js

We also have two new pages: the about page that could be a landing page and the not-found page that is displayed if there was no route found.