Clear the rendering of the navigation bar
The navigation bar is now using a non fluid container for the about and the board page.
I added a link to the board (as well as the brand link)
Finally, I put the menu on the left and the link in white to be consistent with the rest of the menu.