Modularizing the CSS code
The trick is to have a class name for the component and to have all the CSS under that HTML component.
Check the changes in Kanban.css, it says it all.

Also, moving some CSS code from the Kanban container to the Navigation component to have every thing at the right place.

Fixes #3