Implement a draft for the node backend
The node backend is able to answer to HTTP requests such as:
http://localhost:3001/api/lists - GET
http://localhost:3001/api/cards - GET

We have a mongo db where we store all the informations we need for the cards and the lists.
We also have a trello plugin that will fetch Trello API and synchronize our models accordingly.

The goal here is to have a structure flexible enough such that we can add a new integration very easily.

There is certainly some requests that won’t behave as expected, we need to ensure that the create, update, delete and sync works perfectly.
The next step should certainly be to develop unit tests for both our models and the trello plugin.